Update command for helm generator for asset generation to use asset-generation library
