+++ cmd/asset_generation/generate/cmd.go
